requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic~=1.10.4
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.1.0
discord.py==2.2.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
python-dotenv==1.0.0
SQLAlchemy~=2.0.12
typing_extensions==4.5.0
yarl==1.9.1
fluent.runtime==0.4.0